Add files via upload
Helpdesk is an important part of any organization.  In order to allow the organization to work smoothly, we need a system where all the queries or issues raised by users can be resolved. Here, the user should be able to raise a ticket that should be handled by the admin.
So, we on Codebun have develop a Helpdesk Ticket System Project in Java to allow users to raise tickets online. All the activities of raising, the closing tickets will be managed using this application. It has a simple and easy UI also so that users can easily interact with the system.